Add 'zero-dependency' example.
This does not rely on externally defined libraries.
No lib.cxx or use of LIBC. This uses assembler to define a runtime library
which defines a function to print a character to standard output, and defines
an entrypoint that calls the main function with the argument count as its
first argument.